Не работает: регаистрация/авторизация добавление фото с выставкой добавление dog/participant к выбранному Contest Добавлено: список собак у каждой выставки

+++ jsp-demo-tomcat/src/main/java/com/dip/repository/ContestTypeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by moneg on 29.12.2016.
  */
-public interface ContestTypeRepository extends JpaRepository<Contest_type, Long> {
+public interface ContestTypeRepository extends JpaRepository<Contest_type, Integer> {
 }
